feat: add theme switch support
