+++ examples/basic/index.html
@@ -4,9 +4,6 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>React Chess Examples</title>
-    <link rel="stylesheet" href="/node_modules/chessground/assets/chessground.base.css">
-    <link rel="stylesheet" href="/node_modules/chessground/assets/chessground.brown.css">
-    <link rel="stylesheet" href="/node_modules/chessground/assets/chessground.cburnett.css">
   </head>
   <body>
     <div id="root"></div>
