+++ src/types.ts
@@ -3,17 +3,19 @@ import type { Config } from "chessground/config";
 import type { Key } from "chessground/types";
 import { Chess, Move } from "chess.js";
 
-// Pick all properties from HTMLAttributes except 'draggable'
 type HTMLDivPropsWithoutDraggable = Omit<
   React.HTMLAttributes<HTMLDivElement>,
   "draggable"
 >;
+
+export type ChessboardTheme = "brown" | "blue" | "green" | "gray";
 
 export interface ChessboardProps
   extends Partial<Config>,
     HTMLDivPropsWithoutDraggable {
   width?: string | number;
   height?: string | number;
+  theme?: ChessboardTheme;
   fen?: string;
   orientation?: "white" | "black";
   onMove?: (from: Key, to: Key) => void;
@@ -28,4 +30,25 @@ export interface ChessboardProps
 export interface ChessboardRef {
   api: Api | null;
   game: Chess | null;
+}
+
+export interface PGNMetadata {
+  annotations: Record<number, string>;
+  comments: Record<number, string>;
+  variations: Record<number, string[]>;
+  clockTimes: Record<number, string>;
+  evaluations: Record<number, string>;
+  ravs: Record<number, { moves: string; comment?: string }[]>;
+  headers: {
+    White?: string;
+    Black?: string;
+    Date?: string;
+    Event?: string;
+    Site?: string;
+    Result?: string;
+    Round?: string;
+    WhiteElo?: string;
+    BlackElo?: string;
+    ECO?: string;
+  };
 }
